Refactored EditableSelect
